refactor: move functions to separate files
